requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
google-generativeai>=0.2.0
requests==2.31.0
openai==0.27.8
python-multipart==0.0.6
pydantic==2.0.2
httpx==0.24.1